put adc in separate file
